Completed all tests for number 2
For the first function, the . between the variable word and the method trim() was missing so I added that in.

For the second, firstly, the word variable was being used as a string with "" instead of using the passed variable. So I removed those speech marks, also .length is a method and not a function.

For the third function, you was getting the sum of a b and c but not doing anything with it, you was just telling the script to do it while returning nothing. So I combined those two lines so it returns the sum of the three